Update a graph resource as we draw lines
This is a bit of a dead-end and should probably just be rewritten.
I would like to be able to add weights to "edge pairs" so that
pixies can find the optimal path while taking into account having
to slow down for sharp corners.

I think that the only way to accomplish this (with petgraph) is to
add edges inbetween these edge pairs. So rather than reusing nodes
when drawing segments that connect to other nodes, we would give
each segment its own start and end nodes and add edges to connect
them all up.

Do we even need this idea of RoadChunks? Should each line be its
own thing so that every direction change is treated the same way?