Slightly more thematic level names
Had to modify the font to get a lowercase L that was distinguishable
from an uppercase I.

The key to getting FontForge to export with the same general
file size seems to be to ensure that the following things are
checked when generating fonts:

- PS Glyph Names
- Apple

And nothing else. Truetype Hinting *might* less with arrow character
vertical centering.

You can verify that the generated font is generally the same using
fonttools's "ttx" to generate xml that can be diffed.